renamed, Removed Nuget.Core, solution cleaned up
